% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{analyze}
\alias{analyze}
\title{analyze}
\usage{
analyze(
  data,
  participant,
  DV,
  within.vars = NULL,
  between.vars = NULL,
  analysis_type = c("parametric", "non-parametric"),
  posthoc.adj = "bonferroni",
  anova.type = 3,
  transformation = NULL,
  es = c("ges", "pes", "none"),
  ignore_design_not_complete = FALSE
)
}
\arguments{
\item{data}{the data}

\item{participant}{the column name of the participant identifier}

\item{DV}{the column name of the dependent variable}

\item{within.vars}{(optional) a list of within-subject variables}

\item{between.vars}{(optional) a list of between-subject variables}

\item{analysis_type}{one of "parametric" or "non-parametric". Defaults to "parametric"}

\item{posthoc.adj}{the adjustment method for post-hoc tests. Defaults to "bonferroni"}

\item{anova.type}{the type of anova test to perform. Defaults to 3}

\item{transformation}{(optional) a transformation that should be applied to the data before the test.}

\item{es}{the type of effect size to report, one of "ges", "pes" or "none".}

\item{ignore_design_not_complete}{if true, allows to test incomplete designs.}
}
\description{
This function conducts (non-) parametric analyzes of (multi-) factorial experiments
}
