% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.r
\name{ano}
\alias{ano}
\title{ano}
\usage{
ano(
  data,
  dv,
  participantCol,
  within.vars = NULL,
  between.vars = NULL,
  posthoc.adj = "bonferroni",
  anova.type = 2,
  normality.test = shapiro.test,
  latex.digits.round = 2
)
}
\arguments{
\item{data}{the data}

\item{dv}{the name of the column containing the DV}

\item{participantCol}{the column containing an identifier for the participant}

\item{within.vars}{(optional) a vector containing the names of the columns with within-subjects variables}

\item{between.vars}{(optional) a vector containing the names of the columns with between-subjects variables}

\item{posthoc.adj}{(optional) post-hoc adjustment method. Defaults to bonferroni}

\item{anova.type}{(optional) as defined by ezAnova. Defaults to 2.}

\item{normality.test}{(optional) the normality test function to use. Defaults to shapiro.test}

\item{latex.digits.round}{(optional) round the latex output to x digits. Defaults to 2.}
}
\description{
Perform an (RM) anova analysis using ezANOVA including asusmption checks and post-hoc tests
}
