% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{check_complete_design}
\alias{check_complete_design}
\title{check_complete_design}
\usage{
check_complete_design(data, IVs, participantCol, DV)
}
\arguments{
\item{data}{the data}

\item{IVs}{a vector of the names of your IV cols}

\item{participantCol}{the name of your participant ID col}

\item{DV}{the name of the dependent variable column}
}
\description{
This function checks if the data is complete for your design for one DV
}
