% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{data_summary_plot}
\alias{data_summary_plot}
\title{data_summary_plot}
\usage{
data_summary_plot(
  data,
  DV,
  IVs,
  fun.sum = mean,
  fun.error = sd,
  p.scale_fill_manual = NULL,
  p.basesize = 10,
  p.title = NULL,
  theme.fontfamily = NULL,
  theme.fontfamily.device = "win",
  na.rm = FALSE,
  position = "dodge"
)
}
\arguments{
\item{data}{the data}

\item{DV}{the name of the column containing the dependent variable}

\item{IVs}{a vector of the names of the grouping variables}

\item{fun.sum}{(optional) the summary function to use. Defaults to mean.}

\item{fun.error}{(optional) the error function to use. Defaults to sd.}

\item{p.scale_fill_manual}{(optional) a color scale for the plot}

\item{p.basesize}{(optional) font size of the plot}

\item{p.title}{(optinal) a title}

\item{theme.fontfamily}{(optional) a font for your plot. Must be registered using extrafont.}

\item{theme.fontfamily.device}{(optional) the device for extrafont. Defaults to win.}

\item{na.rm}{(default = false) if NAs should be removed before plotting}
}
\description{
This function produces a summary plot of your data
}
