% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{data_summary_plot_all}
\alias{data_summary_plot_all}
\title{data_summary_plot_all}
\usage{
data_summary_plot_all(
  data,
  DVs,
  IVs,
  outdir = file.path(getwd(), "plots"),
  prefix = "plot",
  p.width = NULL,
  p.height = NULL,
  fun.sum = mean,
  fun.error = sd,
  scale_fill_manual = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{data}{the data}

\item{DVs}{vector containing the names of the columns with the dependent variables}

\item{IVs}{a vector of the names of the grouping variables}

\item{outdir}{(optional) a file.path pointing to the directory to save the plots. Defaults to 'plots' within the current working dir.}

\item{prefix}{(optional) a string to prefix the file names of the plots. Defaults to 'plot'}

\item{p.width}{(optional) the width of the plot}

\item{p.height}{(optional) the height of the plot}

\item{fun.sum}{(optional) the summary function to use. Defaults to mean.}

\item{fun.error}{(optional) the error function to use. Defaults to sd.}

\item{na.rm}{if nas shoudl be removed before plotting}

\item{p.scale_fill_manual}{(optional) a color scale for the plot}
}
\description{
This function produces summary plots of multiple dependent variables given a fixed set of grouping variables
}
