% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{data_summary_plot_multiple}
\alias{data_summary_plot_multiple}
\title{data_summary_plot_multiple}
\usage{
data_summary_plot_multiple(
  data,
  DVs,
  IVs,
  outdir = file.path(getwd(), "plots"),
  p.width = NULL,
  p.height = NULL,
  prefix = NULL,
  fun.sum = mean,
  fun.error = sd,
  scale_fill_manual = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{data}{the data}

\item{DVs}{vector containing the names of the columns with the dependent variables}

\item{IVs}{a vector of the names of the grouping variables}

\item{outdir}{a file.path pointing to the directory to save the plots}

\item{p.width}{(optional) the width of the plot}

\item{p.height}{(optional) the height of the plot}

\item{prefix}{(optional) a string to prefix the filenames of the plots}

\item{fun.sum}{(optional) the summary function to use. Defaults to mean.}

\item{fun.error}{(optional) the error function to use. Defaults to sd.}

\item{na.rm}{(default = false) if nas should be removed before plotting}

\item{p.scale_fill_manual}{(optional) a color scale for the plot}
}
\description{
This function produces summary plots of multiple dependent variables given a fixed set of grouping variables
}
