% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{descriptives}
\alias{descriptives}
\title{descriptives}
\usage{
descriptives(data, DV, IVs)
}
\arguments{
\item{data}{the data}

\item{DV}{the dependent variable}

\item{IVs}{a list of independent variables}
}
\description{
calculate descriptives for all subsets
}
