% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{grid_arrange_caption}
\alias{grid_arrange_caption}
\title{grid_arrange_caption}
\usage{
grid_arrange_caption(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right")
)
}
\arguments{
\item{nrow}{(optional) number of rows (defaults to 1)}

\item{position}{(optional) position of the legend, defaults to bottom}
}
\description{
Arranges the caption when doing multiple plots in one figure. Not finished, use at own risk.
}
