% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{grid_arrange_shared_legend}
\usage{
grid_arrange_shared_legend(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right")
)
}
\arguments{
\item{nrow}{(optional) number of rows (defaults to 1)}

\item{position}{(optional) position of the legend, defaults to bottom}
}
\description{
Arranges multiple ggplot grobs in one plot
}
