% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_analyze.R
\name{lazy_analyze}
\alias{lazy_analyze}
\title{lazy_analyze}
\usage{
lazy_analyze(
  lazy_model,
  dv,
  analysis_type = c("aov", "art", "lme", "glme", "friedman"),
  posthoc.adj = "bonf",
  anova.type = 3,
  transformation = NULL,
  family = c(NULL, "poisson", "binomial"),
  nAGQ = NULL,
  lme.formula = NULL
)
}
\arguments{
\item{lazy_model}{the data model}

\item{dv}{the column name of the dependent variable}

\item{analysis_type}{one of "aov", "art", "friedman", "lme", "glme".}

\item{posthoc.adj}{the adjustment method for post-hoc tests. Defaults to "bonf"}

\item{anova.type}{the type of anova test to perform. Defaults to 3}

\item{transformation}{(optional) a transformation that should be applied to the data before the test.}

\item{family}{the family to use for lme or glme.One of "poisson" or "binomial"}

\item{nAGQ}{the nAGQ argument for glmer}

\item{lme.formula}{(optional) a formula for glme or lme calls to override the calculated formula.}
}
\description{
This function conducts (non-) parametric analyzes of (multi-) factorial experiments
}
