% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_analyze.R
\name{lazy_analyze2}
\alias{lazy_analyze2}
\title{lazy_analyze2}
\usage{
lazy_analyze2(
  lazy_model,
  dv,
  analysis_type = c("aov", "art", "lme", "glme", "friedman"),
  posthoc.adj = "bonf",
  anova.type = 3,
  transformation = NULL,
  family = c("poisson", "binominal"),
  nAGQ = NULL
)
}
\arguments{
\item{lazy_model}{the data model}

\item{dv}{the column name of the dependent variable}

\item{analysis_type}{one of "aov", "art", "friedman", "lme", "glme".}

\item{posthoc.adj}{the adjustment method for post-hoc tests. Defaults to "bonferroni"}

\item{anova.type}{the type of anova test to perform. Defaults to 3}

\item{transformation}{(optional) a transformation that should be applied to the data before the test.}
}
\description{
This function conducts (non-) parametric analyzes of (multi-) factorial experiments
}
