% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_plot.R
\name{lazy_arrange_plots}
\alias{lazy_arrange_plots}
\title{lazy_arrange_plots}
\usage{
lazy_arrange_plots(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right")
)
}
\arguments{
\item{nrow}{(optional) number of rows (defaults to 1)}

\item{position}{(optional) position of the legend, defaults to bottom}
}
\description{
Arranges multiple ggplot grobs in one plot
}
