% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_helpers.R
\name{lazy_check_complete_design}
\alias{lazy_check_complete_design}
\title{lazy_check_complete_design}
\usage{
lazy_check_complete_design(lazy_model, dv)
}
\arguments{
\item{lazy_model}{the data model}

\item{dv}{the name of the dependent variable column}
}
\description{
This function checks if the data is complete for your design for one DV
}
