% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_descriptives.R
\name{lazy_descriptives}
\alias{lazy_descriptives}
\title{descriptives}
\usage{
lazy_descriptives(lazy_model, dv, ivs = NULL)
}
\arguments{
\item{lazy_model}{the data in tidy format}

\item{dv}{the name of the column containing the dependent variable}

\item{ivs}{(optional) a vector of strings containing names of the columns identifying the independent variables to plot}
}
\description{
calculate descriptives for all subsets
}
