% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_helpers.R
\name{lazy_import_google_form}
\alias{lazy_import_google_form}
\title{import_google_form}
\usage{
lazy_import_google_form(
  url,
  pre_questions,
  post_questions,
  nr_of_conditions,
  participant_column,
  verbose = FALSE
)
}
\arguments{
\item{url}{the spredsheet url}

\item{pre_questions}{the number of pre question columns}

\item{post_questions}{the number of post question columns}

\item{nr_of_conditions}{the number of conditions}

\item{participant_column}{the column number containing the participant id.}

\item{x}{The data}
}
\description{
Imports data from google forms (sheets). Expects study data to be in a specific format:
}
\details{
0:n pre questions (demographics etc.)
0:n questions per condition for 0:n conditions
0:n post study questions
}
