% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_load.R
\name{lazy_load}
\alias{lazy_load}
\title{lazy_load}
\usage{
lazy_load(pattern, sep = "/", read = readr::read_csv)
}
\arguments{
\item{pattern}{pattern describing the path to the files and variable names to extract from the path. Variables should be encapsulated in curly brackets.}

\item{sep}{path seperator. Defaults to "/"}

\item{read}{function used for reading the files. defaults to readr::read_csv}
}
\description{
loads data from multiple files into one data frame.
}
\examples{

\dontrun{
lazy_load("testdata/logs/day{*}/log_study_{participant}_{*}_{TIME}_{*}_{CONDITION}.csv")
}
}
