% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyhci.R
\name{lazy_model}
\alias{lazy_model}
\title{lazy_model}
\usage{
lazy_model(
  data,
  participant,
  within.vars = NULL,
  between.vars = NULL,
  make_factor = FALSE
)
}
\arguments{
\item{data}{the data in tidy format}

\item{participant}{the name of the column containing the participant identifier}

\item{within.vars}{(optional) a vector of strings containing names of the columns identifying within independent variables}

\item{between.vars}{(optional) a vector of strings containing names of the columns identifying within independent variables}

\item{make_factor}{indicates if columns should automatically be converted to factors (default FALSE)}

\item{ivs}{a vector of strings containing names of the columns containing the independent variables}
}
\description{
This function produces a lazy mode
}
