% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_plot.R
\name{lazy_plot}
\alias{lazy_plot}
\title{lazy_plot}
\usage{
lazy_plot(
  lazy_model,
  dv,
  ivs = NULL,
  fun.sum = mean,
  fun.error = sd,
  p.scale_fill_manual = NULL,
  p.basesize = 10,
  p.title = NULL,
  theme.fontfamily = NULL,
  na.rm = FALSE,
  position = "dodge",
  collapse_trials = FALSE
)
}
\arguments{
\item{lazy_model}{the data in tidy format}

\item{dv}{the name of the column containing the dependent variable}

\item{ivs}{(optional) a vector of strings containing names of the columns identifying the independent variables to plot}

\item{fun.sum}{(optional) the summary function to use. Defaults to mean.}

\item{fun.error}{(optional) the error function to use. Defaults to sd.}

\item{p.scale_fill_manual}{(optional) a color scale for the plot}

\item{p.basesize}{(optional) font size of the plot}

\item{p.title}{(optinal) a title}

\item{theme.fontfamily}{(optional) a font for your plot. Must be registered using extrafont.}

\item{na.rm}{(default = false) if NAs should be removed before plotting}

\item{position}{= "dodge" (default = dodge) if NAs should be removed before plotting}

\item{collapse_trials}{(default = false) if trials should be averaged per participant before plotting}
}
\description{
This function provides means for plotting lazy models
}
