% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_plot.R
\name{lazy_plot_likert}
\alias{lazy_plot_likert}
\title{likert_plot_model}
\usage{
lazy_plot_likert(
  lazy_model,
  dv,
  grouping = NULL,
  drop = NULL,
  survey_vector = c("auto", "likert-4", "likert-5", "likert-6", "likert-7", "manual",
    "as_is"),
  survey_vector_manual = NULL,
  title = DV.pretty,
  yblank = FALSE,
  ordered = FALSE,
  colorscale = NULL,
  p.basesize = 10,
  percentagelabel = FALSE,
  theme.fontfamily = NULL,
  wrap = 1000,
  group.order = NULL
)
}
\arguments{
\item{lazy_model}{the lazy model}

\item{dv}{the name of the column containing the dependent variable.}

\item{grouping}{(optional) one or more grouping IVs}

\item{drop}{(optional) one or more of the IVs that should be dropped before plotting.}

\item{survey_vector}{one of auto, likert-4, likert-5,likert-6, likert-7, as_is, or manual. If manual, you need to pass a survey_vector_manual}

\item{survey_vector_manual}{if 'survey_vector' is 'manual', pass a fitting survey vector here}

\item{title}{(optional) the title of theplot}

\item{yblank}{(optional) if the y Axis should be blank.}

\item{ordered}{(optional)if the plot should be ordered.}

\item{colorscale}{(optional) the colorscale to use.}

\item{p.basesize}{(optional) that base size of the plot}

\item{percentagelabel}{(optional) if the percentage labels should be shown}

\item{theme.fontfamily}{(optional) the font family to use}

\item{wrap}{description}

\item{group.order}{(optional) the group.order. If grouped, the order of the grouping.}
}
\description{
This function produces a likert model that can be plotted
}
