% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{likert_plot_model}
\alias{likert_plot_model}
\title{likert_plot_model}
\usage{
likert_plot_model(
  data,
  DV,
  IVs,
  participantCol,
  grouping = NULL,
  drop = NULL,
  survey_vector = c("auto", "likert-4", "likert-5", "likert-6", "likert-7", "manual"),
  survey_vector_manual = NULL
)
}
\arguments{
\item{data}{the data}

\item{DV}{the name of the column containing the dependent variable.}

\item{IVs}{the names of the columns containing the IVs. The IVs should be numbers representing the likert items.}

\item{participantCol}{the name of the column containing the participant ID}

\item{grouping}{(optional) one or more grouping IVs}

\item{drop}{(optional) one or more of the IVs that should be dropped before plotting.}

\item{survey_vector}{one of auto, likert-4, likert-5,likert-6, likert-7 or manual. If manual, you need to pass a survey_vector_manual}

\item{survey_vector_manual}{if 'survey_vector' is 'manual', pass a fitting survey vector here}
}
\description{
This function produces a likert model that can be plotted
}
\seealso{
\code{\link{plot_likert}}
}
