% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_likert}
\alias{plot_likert}
\title{plot_likert}
\usage{
plot_likert(
  likert.model,
  title = "",
  yblank = FALSE,
  ordered = FALSE,
  colorscale = NULL,
  p.basesize = 19,
  percentagelabel = FALSE,
  theme.fontfamily = NULL,
  theme.fontfamily.device = "win"
)
}
\arguments{
\item{likert.model}{a model generated by \code{\link{likert_plot_model}}}

\item{title}{the title of theplot}

\item{yblank}{(optional) if the y Axis should be blank.}

\item{ordered}{(optional)if the plot should be ordered.}

\item{colorscale}{(optional) the colorscale to use.}

\item{p.basesize}{(optional) that base size of the plot}

\item{percentagelabel}{(optional) if the percentage labels should be shown}

\item{theme.fontfamily}{(optional) the font family to use}

\item{theme.fontfamily.device}{(optional) device to use for plotting. defaults to win}
}
\description{
This function produces a likert plot
}
\seealso{
\code{\link{likert_plot_model}}
}
