% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_helpers.R
\name{se}
\alias{se}
\title{se}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{The data}

\item{na.rm}{remove NA (defaults to FALSE)}
}
\description{
This function allows you to calculate the standard error
}
