% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{set_col_type}
\alias{set_col_type}
\title{set_col_type}
\usage{
set_col_type(x, cols, coltype = c("numeric", "character", "factor"))
}
\arguments{
\item{x}{The data}

\item{cols}{a vector of col numbers}

\item{coltype}{one of 'numeric', 'character' 'factor'}
}
\description{
Batch-sets column names
}
