% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_new.r
\name{lazy_analyze}
\alias{lazy_analyze}
\title{lazy_analyze}
\usage{
lazy_analyze(
  data,
  participant,
  DV,
  within.vars = NULL,
  between.vars = NULL,
  analysis_type = c("aov", "art", "lme", "glme", "friedman"),
  posthoc.adj = "bonf",
  anova.type = 3,
  transformation = NULL,
  family = c("poisson", "binominal"),
  nAGQ = NULL,
  make_factor = FALSE
)
}
\arguments{
\item{data}{the data}

\item{participant}{the column name of the participant identifier}

\item{DV}{the column name of the dependent variable}

\item{within.vars}{(optional) a list of within-subject variables}

\item{between.vars}{(optional) a list of between-subject variables}

\item{analysis_type}{one of "aov", "art", "friedman", "lme", "glme".}

\item{posthoc.adj}{the adjustment method for post-hoc tests. Defaults to "bonferroni"}

\item{anova.type}{the type of anova test to perform. Defaults to 3}

\item{transformation}{(optional) a transformation that should be applied to the data before the test.}
}
\description{
This function conducts (non-) parametric analyzes of (multi-) factorial experiments
}
