% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_latex.R
\name{lazy_latex}
\alias{lazy_latex}
\title{descriptives}
\usage{
lazy_latex(lazy_object)
}
\arguments{
\item{lazy_model}{an object produced by lazy_hci.}
}
\description{
provides latex code for lazy_objects.
}
\details{
Currently supported: lazy_descriptives (table)
}
